Use HttpClientResponseHandler object instead of inline. Use javax by default in C generator
