+++ compiler/cpp/src/thrift/generate/t_java_generator.cc
@@ -80,7 +80,7 @@ public:
     rethrow_unhandled_exceptions_ = false;
     unsafe_binaries_ = false;
     annotations_as_metadata_ = false;
-    javax_annotations_ = false;
+    jakarta_annotations_ = false;
     for (iter = parsed_options.begin(); iter != parsed_options.end(); ++iter) {
       if (iter->first.compare("beans") == 0) {
         bean_style_ = true;
@@ -129,8 +129,8 @@ public:
         unsafe_binaries_ = true;
       } else if (iter->first.compare("annotations_as_metadata") == 0) {
         annotations_as_metadata_ = true;
-      } else if (iter->first.compare("javax_annotations") == 0) {
-        javax_annotations_ = true;
+      } else if (iter->first.compare("jakarta_annotations") == 0) {
+        jakarta_annotations_ = true;
       } else {
         throw "unknown option java:" + iter->first;
       }
@@ -468,7 +468,7 @@ private:
   bool rethrow_unhandled_exceptions_;
   bool unsafe_binaries_;
   bool annotations_as_metadata_;
-  bool javax_annotations_;
+  bool jakarta_annotations_;
 };
 
 /**
@@ -5811,10 +5811,10 @@ void t_java_generator::generate_java_scheme_lookup(ostream& out) {
 void t_java_generator::generate_javax_generated_annotation(ostream& out) {
   time_t seconds = time(nullptr);
   struct tm* now = localtime(&seconds);
-  if (javax_annotations_) {
-    indent(out) << "@javax.annotation.Generated(value = \"" << autogen_summary() << "\"";
-  } else {
+  if (jakarta_annotations_) {
     indent(out) << "@jakarta.annotation.Generated(value = \"" << autogen_summary() << "\"";
+  } else {
+    indent(out) << "@javax.annotation.Generated(value = \"" << autogen_summary() << "\"";
   }
 
   if (undated_generated_annotations_) {
@@ -5857,6 +5857,6 @@ THRIFT_REGISTER_GENERATOR(
     "                     undated: suppress the date at @Generated annotations\n"
     "                     suppress: suppress @Generated annotations entirely\n"
     "    unsafe_binaries: Do not copy ByteBuffers in constructors, getters, and setters.\n"
-    "    javax_annotations: generate javax annotations (jakarta by default)\n"
+    "    jakarta_annotations: generate jakarta annotations (javax by default)\n"
     "    annotations_as_metadata:\n"
     "                     Include Thrift field annotations as metadata in the generated code.\n")
